fix: merge player into users to map closer to lichess-api docs
