+++ src/api/mod.rs
@@ -7,7 +7,6 @@ pub mod external_engine;
 pub mod games;
 pub mod messaging;
 pub mod openings;
-pub mod player;
 pub mod puzzles;
 pub mod studies;
 pub mod tablebase;
