+++ src/api/users.rs
@@ -83,4 +83,15 @@ impl LichessApi<reqwest::Client> {
     ) -> Result<crate::model::Ok> {
         self.get_single_model(request.into()).await
     }
+
+    pub async fn get_all_top_10(&self) -> Result<Leaderboards> {
+        self.get_single_model(top_10::GetRequest::default()).await
+    }
+
+    pub async fn get_one_leaderboard(
+        &self,
+        request: impl Into<leaderboard::GetRequest>,
+    ) -> Result<Leaderboard> {
+        self.get_single_model(request.into()).await
+    }
 }
