+++ src/model/users/mod.rs
@@ -3,12 +3,14 @@ pub mod autocomplete;
 pub mod autocomplete_name;
 pub mod by_id;
 pub mod crosstable;
+pub mod leaderboard;
 pub mod live_streamers;
 pub mod note;
 pub mod performance;
 pub mod public;
 pub mod rating_history;
 pub mod status;
+pub mod top_10;
 
 use std::collections::HashMap;
 
@@ -173,4 +175,92 @@ pub struct Note {
     to: LightUser,
     text: String,
     date: u64,
+}
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+pub struct LeaderboardPerf {
+    pub rating: u32,
+    pub progress: i32,
+}
+
+// Lichess returns { perfs: { "{PerfType}": { rating: u32, progress: u32} } },
+// this therfore more accurate than a struct with `variant_name: Perf` for each variant.
+#[skip_serializing_none]
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub enum LeaderboardPerfs {
+    Bullet(LeaderboardPerf),
+    Blitz(LeaderboardPerf),
+    Rapid(LeaderboardPerf),
+    Classical(LeaderboardPerf),
+    UltraBullet(LeaderboardPerf),
+    Chess960(LeaderboardPerf),
+    Crazyhouse(LeaderboardPerf),
+    Antichess(LeaderboardPerf),
+    Atomic(LeaderboardPerf),
+    Horde(LeaderboardPerf),
+    KingOfTheHill(LeaderboardPerf),
+    RacingKings(LeaderboardPerf),
+    ThreeCheck(LeaderboardPerf),
+}
+
+impl LeaderboardPerfs {
+    pub fn into_perf(self) -> LeaderboardPerf {
+        match self {
+            Self::Bullet(perf)
+            | Self::Blitz(perf)
+            | Self::Rapid(perf)
+            | Self::Classical(perf)
+            | Self::UltraBullet(perf)
+            | Self::Chess960(perf)
+            | Self::Crazyhouse(perf)
+            | Self::Antichess(perf)
+            | Self::Atomic(perf)
+            | Self::Horde(perf)
+            | Self::KingOfTheHill(perf)
+            | Self::RacingKings(perf)
+            | Self::ThreeCheck(perf) => perf,
+        }
+    }
+}
+
+impl From<LeaderboardPerfs> for LeaderboardPerf {
+    fn from(perfs: LeaderboardPerfs) -> LeaderboardPerf {
+        perfs.into_perf()
+    }
+}
+
+#[skip_serializing_none]
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Player {
+    pub id: String,
+    pub username: String,
+    pub perfs: LeaderboardPerfs,
+    pub title: Option<Title>,
+    pub patron: Option<bool>,
+    pub online: Option<bool>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct Leaderboards {
+    pub bullet: Vec<Player>,
+    pub blitz: Vec<Player>,
+    pub rapid: Vec<Player>,
+    pub classical: Vec<Player>,
+    pub ultra_bullet: Vec<Player>,
+    pub chess960: Vec<Player>,
+    pub crazyhouse: Vec<Player>,
+    pub antichess: Vec<Player>,
+    pub atomic: Vec<Player>,
+    pub horde: Vec<Player>,
+    pub king_of_the_hill: Vec<Player>,
+    pub racing_kings: Vec<Player>,
+    pub three_check: Vec<Player>,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Leaderboard {
+    pub users: Vec<Player>,
 }
